% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amtrak.R
\docType{data}
\name{amtrak}
\alias{amtrak}
\title{Amtrak passengers}
\format{
A data frame with 87 observations of 4 variables
\describe{
  \item{month}{Date variable that is the first month of ridership statistics}
  \item{num_passengers}{Number of passengers in thousands}
  \item{t}{time period, from 1-87}
  \item{season}{month}
}
}
\source{
Practical Time Series Forecasting with R, by Galit Shmueli and Kenneth Lichtendahl
}
\usage{
amtrak
}
\description{
Number of monthly passengers in thousands from January 1997 through March 2004.
}
\keyword{datasets}
