% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auswine.R
\docType{data}
\name{auswine}
\alias{auswine}
\title{Monthly Australian Wine Sale}
\format{
A data frame with 187 observations of 7 variables
\describe{
  \item{winet}{month 1-187}
  \item{fortw}{fortified wine}
  \item{dryw}{dry white}
  \item{sweetw}{sweet white}
  \item{red}{red wine}
  \item{rose}{rose wine}
  \item{spark}{sparlking wine}
}
}
\source{
Introductory Time Series with R, by Cowpertwait and Metcalfe
}
\usage{
auswine
}
\description{
Monthly sales of wine (in thousands of liters) in Australia from January 1980 through July 1995.
}
\keyword{datasets}
