% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_share.R
\docType{data}
\name{bike_share}
\alias{bike_share}
\title{Bike share}
\format{

}
\source{
UCI Machine Learning Repository: http://archive.ics.uci.edu/ml/datasets/Bike+Sharing+Dataset#
}
\usage{
bike_share
}
\description{
Number of bike share rides by hour.
}
\details{
From the authors: Bike-sharing rental process is highly correlated to the environmental and seasonal settings. For instance, weather conditions,
precipitation, day of week, season, hour of the day, etc. can affect the rental behaviors. The core data set is related to  
the two-year historical log corresponding to years 2011 and 2012 from Capital Bikeshare system, Washington D.C., USA which is 
publicly available in http://capitalbikeshare.com/system-data.
}
\keyword{datasets}
