% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackberry.R
\docType{data}
\name{blackberry}
\alias{blackberry}
\title{Blackberry Sales Figures}
\format{
A data frame with 4769 observations of 2 variables
\describe{
  \item{quarter}{Quarter of year}
  \item{year}{integer 4-16 indicating 2004-2016}
  \item{handheld}{Blackberry handhed devices sold, in millions.}
  \item{tablets}{Blackberry tablets sold, in millions.}
}
}
\source{
\url{https://barefigur.es/companies/blackberry/products/}
}
\usage{
blackberry
}
\description{
Sales of Blackberry handheld and tabled devices, 2004-2016.
}
\keyword{datasets}
