% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_gas.R
\docType{data}
\name{canada_gas}
\alias{canada_gas}
\title{Gas production in Canada
 
Monthly gas production in Canada.}
\format{
A time series of length 542
}
\source{
expsmooth package
}
\usage{
canada_gas
}
\description{
Gas production in Canada
 
Monthly gas production in Canada.
}
\keyword{datasets}
