% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbe.R
\docType{data}
\name{cbe}
\alias{cbe}
\title{Chocolate, beer and electricity in Australia}
\format{
A data frame with 396 observations of 3 variables
\describe{
  \item{choc}{Amount of chocolate produces (tons)}
  \item{beer}{Amount of beer produced (ml)}
  \item{elec}{Amount fo electricity consumed (kWh)}
}
}
\source{
Introductory Time Series with R, by Cowpertwait and Metcalf
}
\usage{
cbe
}
\description{
Monthly production of three items in Australia from 1958-1990. Data starts in January, 1958 and runs through December 1990.
}
\keyword{datasets}
