% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\docType{data}
\name{cci}
\alias{cci}
\title{Consumer Confidence Index}
\format{
A vector containing 41 observations
}
\source{
`expsmooth` package
}
\usage{
cci
}
\description{
Monthly US consumer confidence index from June 1997-October 2000
}
\keyword{datasets}
