% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denver.R
\docType{data}
\name{denver}
\alias{denver}
\title{Transit boardings in Denver}
\format{
A time series of length 68
}
\source{
TSA package
}
\usage{
denver
}
\description{
Log of number of transit boardings in Denver, CO.
}
\keyword{datasets}
