% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing.R
\docType{data}
\name{housing}
\alias{housing}
\title{Australian Housing Starts and Activity}
\format{
A data frame with 4769 observations of 2 variables
\describe{
  \item{year}{From 1996-2006.}
  \item{quarter}{Integer value from 1-4}
  \item{approvals}{Total dwellings approved per month averaged over the quarter}
  \item{activity}{Chain volume measured in millions of 2004 AUD}
}
}
\source{
Inroductory Time Series with R, by Cowpertwait and Metcalfe
}
\usage{
housing
}
\description{
From the Australian Bureau of Statistics. Building approvals for each month are published as well
as the value of building that was done each quarter. The building approvals are averaged to give quarterly
values.
}
\keyword{datasets}
