% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipod.R
\docType{data}
\name{ipod}
\alias{ipod}
\title{Quarterly sales of ipods}
\format{
A data frame with 52 observations of 3 variable2
\describe{
  \item{category}{quarter}
  \item{i_pod}{number of sales of ipods, in millions}
  \item{yoy_growth}{year over year growth, percent}
}
}
\source{
https://barefigur.es/companies/apple/ipod/
}
\usage{
ipod
}
\description{
Quarterly sales of ipods
}
\keyword{datasets}
