% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maine.R
\docType{data}
\name{maine}
\alias{maine}
\title{Maine unemployment data}
\format{
A data frame with 128 observations of 1 variable
\describe{
  \item{unemploy}{Unemployment rate}
}
}
\source{
Introductory Time Series with R by Cowpertwait and Metcalfe
}
\usage{
maine
}
\description{
Monthly unemployment figures in Maine starting from January, 1996.
}
\keyword{datasets}
