% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_temperature.R
\docType{data}
\name{monthly}
\alias{monthly}
\title{Global Temperatures}
\format{
A data frame with 3288 observations of 3 variables
\describe{
  \item{source}{Either GCAG (global Climate at a Glance) or GISTEMP (GISS Surface Temperature)}
  \item{Date}{End date of the month over which the average temperature is taken}
  \item{mean}{Mean global temperature}
}
}
\source{
\url{https://datahub.io/core/global-temp}
}
\usage{
monthly
}
\description{
A time series of mean global temperatures in terms of departure from overall mean
from 1880 through 2016.
}
\keyword{datasets}
