% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motororg.R
\docType{data}
\name{motororg}
\alias{motororg}
\title{Complaints to a motoring org}
\format{
A data frame with 48 observations of 1 variable
\describe{
  \item{complaints}{number of complaints}
}
}
\source{
Introductory Time Series with R, by Cowpertwait and Metcalfe
}
\usage{
motororg
}
\description{
Number of monthly complaints to a "motoring org" from 1996-1999
}
\keyword{datasets}
