% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pounds_nz.R
\docType{data}
\name{pounds_nz}
\alias{pounds_nz}
\title{Pound NZ Dollar Exchange Rate Data}
\format{
A data frame with 39 observations of 1 variable
\describe{
  \item{xrate}{Exchange rate}
}
}
\source{
Introductory Time Series with R, by Cowpertwait and Metcalfe
}
\usage{
pounds_nz
}
\description{
Quarterly mean exchange rate between pounds and NZD.
}
\keyword{datasets}
