% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccines.R
\docType{data}
\name{vaccines}
\alias{vaccines}
\title{Missouri Covid Vaccines}
\format{
A data frame with changing number of observations of 2 variables
\describe{
  \item{date}{Date of vaccine}
  \item{doses}{Number of vaccines}
}
}
\source{
\url{https://barefigur.es/companies/blackberry/products/}
}
\usage{
vaccines
}
\description{
Number of doses of COVID vaccines given in Missouri by date. This data set will be
updated semi-regularly.
}
\keyword{datasets}
