% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_temperature.R
\docType{data}
\name{yearly}
\alias{yearly}
\title{Global Temperatures}
\format{
A data frame with 274 observations of 3 variables
\describe{
  \item{source}{Either GCAG (global Climate at a Glance) or GISTEMP (GISS Surface Temperature)}
  \item{year}{Year, from 1880-2016}
  \item{mean}{Mean global temperature}
}
}
\source{
\url{https://za.investing.com/}
}
\usage{
yearly
}
\description{
A time series of mean global temperatures in terms of departure from overall mean
from 1880 through 2016.
}
\keyword{datasets}
