% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\alias{sim_data_1}
\alias{sim_data_2}
\title{Simulated time series data}
\format{
Vectors with varying lengths

An object of class \code{ts} of length 300.
}
\source{
Constructed by Darrin Speegle
}
\usage{
sim_data_1

sim_data_2
}
\description{
The data sim_data_*, where * is an integer, are simulated time series of various types.
}
\keyword{datasets}
